+++ tclpkg/gv/gv_tcl_init.c
@@ -19,7 +19,7 @@
 static size_t gv_string_writer(GVJ_t *job, const char *s, size_t len)
 {
   // clamp to INT_MAX
-  int l = len > (size_t)INT_MAX ? INT_MAX : (int)len;
+  int l = len > INT_MAX ? INT_MAX : (int)len;
   Tcl_AppendToObj((Tcl_Obj*)(job->output_file), s, l);
   return (size_t)l;
 }
@@ -27,7 +27,7 @@ static size_t gv_string_writer(GVJ_t *job, const char *s, size_t len)
 static size_t gv_channel_writer(GVJ_t *job, const char *s, size_t len)
 {
   // clamp to INT_MAX
-  int l = len > (size_t)INT_MAX ? INT_MAX : (int)len;
+  int l = len > INT_MAX ? INT_MAX : (int)len;
   return (size_t)Tcl_Write((Tcl_Channel)(job->output_file), s, l);
 }
